requirements.txt
discord.py
requests
flask
transformers
torch